Make sentrymodifier a robolectric test
